# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing TV series catalog",
        "title": "Serial Catalog API",
        "contact": {
            "email": "evgenij.bx@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/otus.account.delete/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete a account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Delete account",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.DeleteResponse"
                        }
                    },
                    "400": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/otus.account.get/{id}": {
            "get": {
                "description": "Get detailed information about a account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Get account by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved account",
                        "schema": {
                            "$ref": "#/definitions/handler.AccountItemResponse"
                        }
                    },
                    "400": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/otus.account.login": {
            "post": {
                "description": "Get token by UserName",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Get token by UserName",
                "parameters": [
                    {
                        "description": "Account data",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.AccountLoginRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/otus.account.register": {
            "post": {
                "description": "Add a new account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Register new account",
                "parameters": [
                    {
                        "description": "Account data",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.AccountRegisterRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/otus.season.add": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add a new season to the catalog",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "seasons"
                ],
                "summary": "Create new TV season",
                "parameters": [
                    {
                        "description": "Season data",
                        "name": "season",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/catalog.Season"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SeasonItemResponse"
                        }
                    }
                }
            }
        },
        "/otus.season.delete/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete a TV season from catalog",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "seasons"
                ],
                "summary": "Delete season",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Season ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.DeleteResponse"
                        }
                    },
                    "400": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/otus.season.get/{id}": {
            "get": {
                "description": "Get detailed information about a TV season",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "seasons"
                ],
                "summary": "Get season by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Season ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved season",
                        "schema": {
                            "$ref": "#/definitions/handler.SeasonItemResponse"
                        }
                    },
                    "400": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/otus.season.update/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update existing TV season",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "seasons"
                ],
                "summary": "Update season",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Season ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update data",
                        "name": "season",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/catalog.Season"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SeasonItemResponse"
                        }
                    },
                    "400": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "catalog.Season": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "created_by": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "moderated": {
                    "type": "boolean"
                },
                "serial_id": {
                    "type": "integer"
                },
                "sort": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "handler.AccountItemResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "object",
                    "properties": {
                        "item": {
                            "$ref": "#/definitions/model.Account"
                        }
                    }
                }
            }
        },
...
        "handler.AccountLoginRequest": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string",
                    "example": "ekobzar"
                },
                "password": {
                    "type": "string",
                    "example": "123456"
                }
            }
        },
        "handler.AccountRegisterRequest": {
            "type": "object",
            "properties": {
                "first_name": {
                    "type": "string",
                    "example": "Kobzar"
                },
                "last_name": {
                    "type": "string",
                    "example": ""
                },
                "login": {
                    "type": "string",
                    "example": "ekobzar"
                },
                "name": {
                    "type": "string",
                    "example": "Evgenij"
                },
                "password": {
                    "type": "string",
                    "example": "123456"
                }
            }
        },
        "handler.DeleteResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "object",
                    "properties": {
                        "deleted": {
                            "type": "boolean"
                        }
                    }
                }
            }
        },
...
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "entity not found"
                }
            }
        },
        "handler.SeasonItemResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "object",
                    "properties": {
                        "item": {
                            "$ref": "#/definitions/catalog.Season"
                        }
                    }
                }
            }
        },
...
        "model.Account": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string",
                    "example": "Kobzar"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "last_name": {
                    "type": "string",
                    "example": ""
                },
                "login": {
                    "type": "string",
                    "example": "ekobzar"
                },
                "name": {
                    "type": "string",
                    "example": "Evgenij"
                },
                "password": {
                    "type": "string",
                    "example": "123456"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}